+++ service/controller/resource/etcdbackup/cleanup.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/giantswarm/apiextensions/v3/pkg/apis/backup/v1alpha1"
+	"github.com/giantswarm/apiextensions-backup/api/v1alpha1"
 	"github.com/giantswarm/microerror"
 )